requirements.txt
google-cloud-storage==3
pycountry==24.6.1
geopy
shapely
loguru
